requirements.txt
backports.csv==1.0.7
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.3
django-appconf==1.0.3
django-crispy-forms==1.7.2
django-cruds-adminlte==0.0.16
django-image-cropping==1.2.0
django-import-export==1.2.0
django-select2==7.1.0
djangoajax==3.1
easy-thumbnails==2.6
et-xmlfile==1.0.1
jdcal==1.4.1
odfpy==1.4.0
openpyxl==2.6.2
Pillow==6.1.0
pytz==2019.1
PyYAML==5.1.1
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
xlrd==1.2.0
xlwt==1.3.0
